pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Olderestin <sdordiuk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "3.14.0"
drf-yasg = "1.21.7"
django-cors-headers = "4.3.1"
djangorestframework-simplejwt = "5.3.1"
pillow = "10.2.0"
psycopg2-binary = "2.9.9"
pydantic-settings = "2.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
